messages: Initialization of uninitialized members varied classes
Fixes the coverity issues:

** 1188168 Uninitialized scalar field
CID 1188168 (#1 of 1): Uninitialized scalar field (UNINIT_CTOR)
2. uninit_member: Non-static class member success is not initialized
in this constructor nor in any functions that it calls.

** 1188169 Uninitialized scalar field
2. uninit_member: Non-static class member map_epoch is not initialized
in this constructor nor in any functions that it calls.
CID 1188169 (#1 of 1): Uninitialized scalar field (UNINIT_CTOR)
4. uninit_member: Non-static class member min_epoch is not initialized
in this constructor nor in any functions that it calls.

** 1188170 Uninitialized scalar field
2. uninit_member: Non-static class member map_epoch is not initialized
in this constructor nor in any functions that it calls.
CID 1188170 (#1 of 1): Uninitialized scalar field (UNINIT_CTOR)
4. uninit_member: Non-static class member min_epoch is not initialized
in this constructor nor in any functions that it calls.

** 1188171 Uninitialized scalar field
2. uninit_member: Non-static class member map_epoch is not initialized
in this constructor nor in any functions that it calls.
CID 1188171 (#1 of 1): Uninitialized scalar field (UNINIT_CTOR)
4. uninit_member: Non-static class member min_epoch is not initialized
in this constructor nor in any functions that it calls.

** 1188172 Uninitialized scalar field
2. uninit_member: Non-static class member map_epoch is not initialized
 in this constructor nor in any functions that it calls.
CID 1188172 (#1 of 1): Uninitialized scalar field (UNINIT_CTOR)
4. uninit_member: Non-static class member min_epoch is not initialized
 in this constructor nor in any functions that it calls.

** 1188173 Uninitialized scalar field
2. uninit_member: Non-static class member map_epoch is not initialized
 in this constructor nor in any functions that it calls.
CID 1188173 (#1 of 1): Uninitialized scalar field (UNINIT_CTOR)
4. uninit_member: Non-static class member min_epoch is not initialized
 in this constructor nor in any functions that it calls.

Signed-off-by: Amit Kumar amitkuma@redhat.com