librbd: Initializing member variables
Fixes the coverity issues:

** 1351727 Uninitialized scalar field
CID 1351727 (#1 of 1): Uninitialized scalar field (UNINIT_CTOR)
2. uninit_member: Non-static class member m_state is not initialized in
this constructor nor in any functions that it calls.

** 1351728 Uninitialized scalar field
CID 1351728 (#1 of 1): Uninitialized scalar field (UNINIT_CTOR)
2. uninit_member: Non-static class member m_state is not initialized in
this constructor nor in any functions that it calls.

** 1351733 Uninitialized scalar field
CID 1351733 (#1 of 1): Uninitialized scalar field (UNINIT_CTOR)
2. uninit_member: Non-static class member snap_id is not initialized in
this constructor nor in any functions that it calls.

Signed-off-by: Amit Kumar amitkuma@redhat.com