rgw: Initialization of member variables done
Fixes the coverity issues:

** 1396118 Uninitialized scalar field
CID 1396118 (#1 of 1): Uninitialized scalar field (UNINIT_CTOR)
2. uninit_member: Non-static class member unix_mode is not initialized
in this constructor nor in any functions that it calls.

** 1396121 Uninitialized scalar field
CID 1396121 (#1 of 1): Uninitialized scalar field (UNINIT_CTOR)
2. uninit_member: Non-static class member is_master is not initialized
in this constructor nor in any functions that it calls.

** 1396137 Uninitialized scalar field
CID 1396137 (#1 of 1): Uninitialized scalar field (UNINIT_CTOR)
3. uninit_member: Non-static class member timer_id is not initialized
in this constructor nor in any functions that it calls.

** 1396138 Uninitialized scalar field
2. uninit_member: Non-static class member zone_short_id is not initialized
in this constructor nor in any functions that it calls.
CID 1396138 (#1 of 1): Uninitialized scalar field (UNINIT_CTOR)
4. uninit_member: Non-static class member pg_ver is not initialized in this
constructor nor in any functions that it calls.

Signed-off-by: Amit Kumar amitkuma@redhat.com