rgw: Intializing is_master, max_aio
Fixes the coverity issue:

** 1396142 Uninitialized scalar field
CID 1396142 (#1 of 1): Uninitialized scalar field (UNINIT_CTOR)
2. uninit_member: Non-static class member max_aio is not initialized in
this constructor nor in any functions that it calls.

** 1396205 Uninitialized scalar field
CID 1396205 (#1 of 1): Uninitialized scalar field (UNINIT_CTOR)
2. uninit_member: Non-static class member is_master is not initialized
in this constructor nor in any functions that it calls.

** 1396150 Uninitialized scalar field
CID 1396150 (#1 of 1): Uninitialized scalar field (UNINIT_CTOR)
2. uninit_member: Non-static class member size is not initialized in
this constructor nor in any functions that it calls.

Signed-off-by: Amit Kumar amitkuma@redhat.com