librbd: Initializing members image,operation,journal
Fixes the coverity issues:

** 1398907 Uninitialized scalar field
CID 1398907 (#1 of 1): Uninitialized scalar field (UNINIT_CTOR)
2. uninit_member: Non-static class member tag_tid is not initialized in
this constructor nor in any functions that it calls.

** 1398906 Uninitialized scalar field
22. uninit_member: Non-static class member m_r_saved is not initialized in
this constructor nor in any functions that it calls.
CID 1398906 (#1 of 1): Uninitialized scalar field (UNINIT_CTOR)
24. uninit_member: Non-static class member m_mirror_mode is not initialized
in this constructor nor in any functions that it calls.

** 1399592 Uninitialized scalar field
2. uninit_member: Non-static class member m_order is not initialized in this
constructor nor in any functions that it calls.
4. uninit_member: Non-static class member m_size is not initialized in this
constructor nor in any functions that it calls.
6. uninit_member: Non-static class member m_features is not initialized in
this constructor nor in any functions that it calls.
8. uninit_member: Non-static class member m_incompatible_features is not
initialized in this constructor nor in any functions that it calls.
10. uninit_member: Non-static class member m_flags is not initialized in
this constructor nor in any functions that it calls.
CID 1399592 (#1-2 of 2): Uninitialized scalar field (UNINIT_CTOR)
12. uninit_member: Non-static class member m_exclusive_locked is not initialized
in this constructor nor in any functions that it calls.

Signed-off-by: Amit Kumar amitkuma@redhat.com