rbd: Initializating unintialized member variables
Fixes the coverity issues:

** 1396107 Uninitialized pointer field
2. uninit_member: Non-static class member on_finish is not initialized
in this constructor nor in any functions that it calls.
CID 1396107 (#1 of 1): Uninitialized pointer field (UNINIT_CTOR)
4. uninit_member: Non-static class member sync_complete is not initialized
in this constructor nor in any functions that it calls.

** 1396113 Uninitialized pointer field
CID 1396113 (#1 of 1): Uninitialized pointer field (UNINIT_CTOR)
2. uninit_member: Non-static class member on_finish is not initialized
in this constructor nor in any functions that it calls.

** 1396114 Uninitialized pointer field
CID 1396114 (#1 of 1): Uninitialized pointer field (UNINIT_CTOR)
2. uninit_member: Non-static class member on_finish is not initialized
in this constructor nor in any functions that it calls.

Signed-off-by: Amit Kumar amitkuma@redhat.com