rbd: Initializing librbd module member variables
** 1396110 Uninitialized pointer field
2. uninit_member: Non-static class member tag_data is not initialized
in this constructor nor in any functions that it calls.
CID 1396110 (#1 of 1): Uninitialized pointer field (UNINIT_CTOR)
4. uninit_member: Non-static class member on_finish is not initialized
in this constructor nor in any functions that it calls.

** 1396135 Uninitialized pointer field
2. uninit_member: Non-static class member m_journaler is not initialized
in this constructor nor in any functions that it calls.
4. uninit_member: Non-static class member m_timer is not initialized in
this constructor nor in any functions that it calls.
6. uninit_member: Non-static class member m_timer_lock is not initialized
in this constructor nor in any functions that it calls.
CID 1396135 (#1 of 1): Uninitialized pointer field (UNINIT_CTOR)
8. uninit_member: Non-static class member m_r_saved is not initialized in
this constructor nor in any functions that it calls.

** 1396136 Uninitialized pointer field
2. uninit_member: Non-static class member m_journaler is not initialized
in this constructor nor in any functions that it calls.
4. uninit_member: Non-static class member m_timer is not initialized in
this constructor nor in any functions that it calls.
6. uninit_member: Non-static class member m_timer_lock is not initialized
in this constructor nor in any functions that it calls.
CID 1396136 (#1 of 1): Uninitialized pointer field (UNINIT_CTOR)
8. uninit_member: Non-static class member m_r_saved is not initialized
in this constructor nor in any functions that it calls.

Signed-off-by: Amit Kumar amitkuma@redhat.com