rbd: Initializing varied members of librbd module
** 1401448 Uninitialized pointer field
CID 1401448 (#1 of 1): Uninitialized pointer field (UNINIT_CTOR)
2. uninit_member: Non-static class member request is not initialized
in this constructor nor in any functions that it calls.

** 1405535 Uninitialized pointer field
5. uninit_member: Non-static class member m_imctx is not initialized
in this constructor nor in any functions that it calls.
7. uninit_member: Non-static class member m_p_features is not initialized
in this constructor nor in any functions that it calls.
9. uninit_member: Non-static class member m_features is not initialized
in this constructor nor in any functions that it calls.
11. uninit_member: Non-static class member m_size is not initialized in
this constructor nor in any functions that it calls.
13. uninit_member: Non-static class member m_is_primary is not initialized
in this constructor nor in any functions that it calls.
CID 1405535 (#1 of 1): Uninitialized pointer field (UNINIT_CTOR)
15. uninit_member: Non-static class member m_force_non_primary is not
initialized in this constructor nor in any functions that it calls.

** 1409840 Uninitialized pointer field
2. uninit_member: Non-static class member m_buf is not initialized
in this constructor nor in any functions that it calls.
CID 1409840 (#1 of 1): Uninitialized pointer field (UNINIT_CTOR)
4. uninit_member: Non-static class member m_pbl is not initialized
in this constructor nor in any functions that it calls.

Signed-off-by: Amit Kumar amitkuma@redhat.com