ceph.spec.in: Include and own /usr/lib64/qemu contents in librbd1
...instead of creating in %post. rpm packages should own all the dirs
they install files into. And there doesn't seem to be any reason to
create the file in %post in the first place.

Like implemented before this change, the affected files fail to
produce meaningful output when rpm -qf is invoked on them, and they
will be left behind when the librbd1 package is removed.

Signed-off-by: Ville Skyttä <ville.skytta@iki.fi>