ceph-disk: Do not start services that are not enabled
Currently, we start the daemons when calling activate from main_trigger.
That makes us discrespect the rules of the init systems.

This fixes up the behaviour by adding --no-start-daemon to the activate
calls and keeping the osd start to the init systems.

If the OSD will be started too early (before ceph-disk completes), the
unit systems will try to restart it after 20 seconds as per our
definitions in the init scripts until it succeeds (or times out).

Signed-off-by: Boris Ranto <branto@redhat.com>