client: allocate and free UserPerm gid lists with malloc/free
Valgrind is unhappy about our turning on supplimentary group handling
with fuse by default. The problem is that we end up calling delete to
free the supplimentary gids list, but fuse uses malloc to allocate it.

We could just convert ceph-fuse to use new/delete instead, but C
programs like ganesha can call ceph_userperm_new() to allocate a new
UserPerm, and they just pass in a pointer to a list of gids, often
malloced.

Convert the UserPerm constructor/destructors to use malloc and free to
handle creating and freeing gid lists. If we end up failing to allocate
a gids list when we need one, just throw a std::bad_alloc exception.

While we're at it, fix ceph_userperm_new to catch that exception instead
of trying to do a std::nothrow exception (as that won't work anyway).

Signed-off-by: Jeff Layton <jlayton@redhat.com>