config: disable skewed utilization warning by default
This has a few problems:

1- It does not do it's analysis over CRUSH rule roots/classes, which
means that an innocent user of classes will see skewed usage (bc hdds are
more full than ssds, say)

2- It does not take degraded clusters into account, which means the warning
will appear when a fresh OSD is added.

See http://tracker.ceph.com/issues/20730

Based on master commit 7832c53 but we've decided not to remove the code
in the older releases so it can be used if appropriate for a
particular cluster configuration.  Also, 7832c53 won't cleanly
cherry-pick so this is easier.

Signed-off-by: David Zafman <dzafman@redhat.com>