rgw: encryption fix the issue when not provide encryption mode
Now, in RGW, if someone want to upload an object using server-side
encryption with providing customer key or kms key id, but not
specify the encryption mode in the
"x-amz-server-side-encryption-customer-algorithm" or
"x-amz-server-side-encryption", the object will be uploaded
successfully without encryption.

This is not a correct way to deal with it. It is better to
return error.

Fixes: http://tracker.ceph.com/issues/21581

Signed-off-by: Enming Zhang <enming.zhang@umcloud.com>
(cherry picked from commit f8c4be8a8943096217d89f7c5abd129fa2414b36)