build/ops: rpm: do not create librbd.so.1 symlink in /usr/lib64/qemu
As reported by Ademar de Souza Reis Jr <areis@redhat.com>:

This symlink should not be necessary anymore. QEMU is properly linked to
ceph/librbd these days (look at the qemu-block-rbd sub-package in Fedora and
the respective package in RHEL).

The symlink was a hack from a time when librbd was distributed and supported by
ceph and the rbd driver would be enabled at runtime only when/if the symlink
was present.

So the right fix is to actually get rid of the symlink altogether and never
touch (much less own) /usr/lib64/qemu.

Signed-off-by: Nathan Cutler <ncutler@suse.com>