os/bluestore/BlueFS: compact log even when sync_metadata sees no work
It's possible that when sync_metadata() is called there won't be any new
log data to flush because it was already flushed for other reasons (e.g.,
because fsync was called).  However, the log may still be large and in
need of compaction.

Signed-off-by: Sage Weil <sage@redhat.com>
(cherry picked from commit dfe5f05f7948da6200229248a41dc1368737380e)

 Conflicts:
	src/os/bluestore/BlueFS.cc
 Some pending release stuff that is now absent at both Luminous and
master.