osd/PGBackend, MOSDPGRecoveryDelete[Reply]: handle incompatible encoding
Handle a pre-luminous encoding without min_epoch for both these
messages.  Bump the header version, so the simple path of inline
decoding works for the future.  For 12.2.0 and pre-luminous, we can
check the SERVER_LUMINOUS flag on the connection to tell which
encoding to use.

This is not a cherry-pick from master since all upgrades must stop at
luminous, and will use the existing encoding before continuing.

Signed-off-by: Josh Durgin <jdurgin@redhat.com>