ceph_mon: use default port when not specified
If the monitor port is not specified in ceph.conf,
it will use 0. However, monmaptool will use
default port in creating monitor map in such case.
The inconsistent behavior is not graceful, and
every time monitor loading, it warns 'mon addr
config option does not match monmap file', which
is confusing.

Signed-off-by: Li Wang <laurence.liwang@gmail.com>