mon: fix dropping mgr metadata for active mgr
drop_standby() was killing it and it was only getting added
back in certain locations.  Instead, make the metadata
drop conditional and only do it in the places we're
really dropping the daemon, not when we're promoting
it to active.

Fixes: http://tracker.ceph.com/issues/21260
Signed-off-by: John Spray <john.spray@redhat.com>
(cherry picked from commit 29c6f9adf178f6611a625740f395e397cad9147b)