msg: Initializing class members in module msg
Fixes the coverity issues:

** 1414530 Uninitialized pointer field
2. uninit_member: Non-static class member tx_cq is not initialized
in this constructor nor in any functions that it calls.
4. uninit_member: Non-static class member rx_cq is not initialized
in this constructor nor in any functions that it calls.
6. uninit_member: Non-static class member tx_cc is not initialized
in this constructor nor in any functions that it calls.
CID 1414530 (#1 of 1): Uninitialized pointer field (UNINIT_CTOR)
8. uninit_member: Non-static class member rx_cc is not initialized
in this constructor nor in any functions that it calls.

** 1416369 Uninitialized pointer field
CID 1416369 (#1 of 1): Uninitialized pointer field (UNINIT_CTOR)
4. uninit_member: Non-static class member send is not initialized
in this constructor nor in any functions that it calls.

** 1405536 Uninitialized pointer field
CID 1405536 (#1 of 1): Uninitialized pointer field (UNINIT_CTOR)
2. uninit_member: Non-static class member msgr is not initialized
in this constructor nor in any functions that it calls.

Signed-off-by: Amit Kumar <amitkuma@redhat.com>