osd/PrimaryLogPG: track last_update, not last_complete, for backfill targets last_complete_ondisk
We trim the PG log based on the min_last_complete_ondisk.  Nobody
else uses the value.

If we have a backfill target, but not backfilling, then the last_complete
on the backfill target(s) do not advance, and our pg log trimming is
effectively stalled indefinitely.  For example, for this pg,

    "up": [
        92,
        18,
        26
    ],
    "acting": [
        92,
        18
    ],
    "backfill_targets": [
        "26"
    ],
    "actingbackfill": [
        "18",
        "26",
        "92"
    ],

in state active+recovery_wait+undersized+degraded+remapped.  The peer
infos are up to date for the backfill target,

            "peer": "26",
            "pgid": "0.36",
            "last_update": "832429'3721336",
            "last_complete": "832429'3721336",

but the PG will not trim because the last_complete on the peer itself is
0'0, and we update peer_last_complete_ondisk[] based on that.

Fix this by updating the peer lcod value to last_complete for acting
peers and last_update for backfill targets.  This makes sense because
last_complete is pretty meaningless for backfill targets since they are
already incomplete and backfilling; it doesn't matter if we trim the pg
log past what they are/were complete to.  All that matters is they persist
the new log entries (hence, use the last_update reported in the write
reply).

Note that this is probably one other source of the "OSDs use more
memory during recovery" meme.

Signed-off-by: Sage Weil <sage@redhat.com>