interval_set: enable interoperability for alt implementations
This allows for the introduction of derived classes that
are optimized for special purposes. Generic iterator and
const_iterator classes allow for interoperation of
iterators between instances of derived classes.

In order to avoid introducing performance regressions in
interval_set, this patch converts only a minimal
selection of methods to virtual methods. In some cases,
template methods are used to implement polymorphic
behavior while maintaining maximal performance.

Signed-off-by: Zac Medico <zmedico@gmail.com>