Add array_interval_set
This interval_set implementation, based on std::vector,
is useful for cases where it provides better performance
than the std::map based implementation. For example,
it can be used to efficiently store a copy or the
result of an intersection operation, while costing
significantly less than it would to perform the same
operations with std::map.

The code is practically identical to interval_set, with
only small modifications in order to adjust to the API
differences between std::vector and std::map.

Signed-off-by: Zac Medico <zmedico@gmail.com>